feat: move Apply and ClearAll button below the filters and change clearAll to link
